+++ rackcity/models/network_port.py
@@ -1,5 +1,6 @@
 from django.db import models
 from .asset import Asset
+from .asset import AssetCP
 
 
 def format_mac_address(mac_address):
@@ -13,16 +14,12 @@ def format_mac_address(mac_address):
     mac_address = mac_address.replace(";", ":")
     # if contains no delimiters:
     if (not mac_address.__contains__(":")):
-        mac_address = ':'.join(a+b for a, b in zip(mac_address[::2], mac_address[1::2]))
+        mac_address = ':'.join(
+            a+b for a, b in zip(mac_address[::2], mac_address[1::2]))
     return mac_address
 
 
-class NetworkPort(models.Model):
-    asset = models.ForeignKey(
-        Asset,
-        on_delete=models.CASCADE,
-        verbose_name="asset",
-    )
+class AbstractNetworkPort(models.Model):
     port_name = models.CharField(
         max_length=150
         # electing not to validate this as it is not user entered
@@ -75,6 +72,19 @@ class NetworkPort(models.Model):
             self.save()
             destination_port.save()
 
+    def save(self, *args, **kwargs):
+        if self.mac_address is not None:
+            self.mac_address = format_mac_address(self.mac_address)
+        super(NetworkPort, self).save(*args, **kwargs)
+
+
+class NetworkPort(AbstractNetworkPort):
+    asset = models.ForeignKey(
+        Asset,
+        on_delete=models.CASCADE,
+        verbose_name="asset",
+    )
+
     class Meta:
         constraints = [
             models.UniqueConstraint(
@@ -83,7 +93,18 @@ class NetworkPort(models.Model):
             ),
         ]
 
-    def save(self, *args, **kwargs):
-        if self.mac_address is not None:
-            self.mac_address = format_mac_address(self.mac_address)
-        super(NetworkPort, self).save(*args, **kwargs)
+
+class NetworkPortCP(AbstractNetworkPort):
+    asset = models.ForeignKey(
+        AssetCP,
+        on_delete=models.CASCADE,
+        verbose_name="asset",
+    )
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                fields=["asset", "port_name"],
+                name="unique network port names on assets",
+            ),
+        ]
