+++ rackcity/models/power_port.py
@@ -1,14 +1,9 @@
 from django.db import models
-from .asset import Asset
+from .asset import Asset, AssetCP
 from .pdu_port import PDUPort
 
 
-class PowerPort(models.Model):
-    asset = models.ForeignKey(
-        Asset,
-        on_delete=models.CASCADE,
-        verbose_name="asset",
-    )
+class AbstractPowerPort(models.Model):
     port_name = models.CharField(
         max_length=150
         # electing not to validate this as it is not user entered
@@ -18,6 +13,29 @@ class PowerPort(models.Model):
         on_delete=models.CASCADE,
         null=True,
         blank=True,
+    )
+
+
+class PowerPort(AbstractPowerPort):
+    asset = models.ForeignKey(
+        Asset,
+        on_delete=models.CASCADE,
+        verbose_name="asset",
+    )
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                fields=['asset', 'port_name'],
+                name='unique power port names on assets'),
+        ]
+
+
+class PowerPortCP(AbstractPowerPort):
+    asset = models.ForeignKey(
+        AssetCP,
+        on_delete=models.CASCADE,
+        verbose_name="asset",
     )
 
     class Meta:
